pyproject.toml
[project]
name = "wikiplot-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "opensearch-py>=3.1.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=3.0,<5.0",
    "numpy<2.0",
    "torch>=2.0,<2.3",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "pydantic>=2.12.5",
    "langsmith>=0.7.30",
]
